Appplied clang-tidy rules to manifest files changed by ingestsql

+++ cvmfs/bigvector.h
@@ -123,7 +123,7 @@ class BigVector {
 
  private:
   static const size_t kNumInit = 16;
-  static const size_t kMmapThreshold = 128*1024;
+  static const size_t kMmapThreshold = static_cast<int64_t>(128*1024);
 
   Item *Alloc(const size_t num_elements) {
     Item *result;
